backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.0.0
langchain==0.3.19
langchain-openai==0.3.7
langchain-community==0.3.18
langchain-core==0.3.40
supabase==2.13.0
requests==2.32.3
python-multipart==0.0.6